+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-0.6.0
+0.6.0 (work in progress)
 =====
-- Support for Nextcloud and Wordpress (work in progress)
+- Apache support (only on Debian for now)
+- Support for Nextcloud and Wordpress
+- Redis support
+- IPv6-Support
 
 0.5.0
 =====